Unified i64 tags for all ADTs (flat struct + heap nodes)
Previously: non-recursive ADTs used i64 tags, recursive ADTs used i8.
Now both use i64 consistently:
- compiled_runtime.c: yona_rt_adt_alloc/get_tag use int64_t
- LLVM declarations: rt_adt_alloc(i64, i64), rt_adt_get_tag returns i64
- No more i8-to-i64 zext workarounds in pattern matching

Also: proper field type tracking for recursive ADT pattern matching.
Fields are typed from AdtInfo.field_types instead of heuristic.
Pointer-based types (FUNCTION, SEQ, STRING, ADT) get inttoptr cast.
This enables thunk fields in ADTs to be callable.

+++ src/compiled_runtime.c
@@ -166,14 +166,14 @@ void yona_rt_print_dict(int64_t* dict) {
 /* Heap-allocated ADT nodes: [tag (i8), field0 (i64), field1 (i64), ...] */
 /* Used for recursive types like List a = Cons a (List a) | Nil        */
 
-void* yona_rt_adt_alloc(int8_t tag, int64_t num_fields) {
+void* yona_rt_adt_alloc(int64_t tag, int64_t num_fields) {
     int64_t* node = (int64_t*)malloc((1 + num_fields) * sizeof(int64_t));
-    node[0] = (int64_t)tag;
+    node[0] = tag;
     return node;
 }
 
-int8_t yona_rt_adt_get_tag(void* node) {
-    return (int8_t)(((int64_t*)node)[0]);
+int64_t yona_rt_adt_get_tag(void* node) {
+    return ((int64_t*)node)[0];
 }
 
 int64_t yona_rt_adt_get_field(void* node, int64_t index) {
